Crate typed_headers

Typed HTTP header serialization and deserialization.

This crate is still in its early, experimental stages. It currently takes a fairly pedantic view of parsing, and tries to support exactly what’s specified in the HTTP RFCs.

The HeaderMapExt extension trait provides new methods on the http::HeaderMap type to insert, retrieve, and remove headers in a typed manner.

Re-exports

pub extern crate http;

pub extern crate mime;

Modules

util

Structs

Accept

Accept header, defined in RFC7231

AcceptEncoding

Accept-Encoding header, defined in RFC7231

Allow

Allow header, defined in RFC7231

AuthScheme

An authorization scheme.

Authorization

Authorization header, defined in RFC7235

ContentCoding

A content coding, used in the Accept-Encoding and Content-Encoding headers.

ContentEncoding

Content-Encoding header, defined in RFC7231

ContentLength

Content-Length header, defined in RFC7230

ContentType

Content-Type header, defined in RFC7231

Credentials

Authentication credentials, as described in RFC7235.

Error

An error serializing or deserializing a header.

Host

The Host header, defined in RFC7230.

HttpDate

ProxyAuthorization

Proxy-Authorization header, defined in RFC7235

Quality

A quality value, as specified in RFC7231.

QualityItem

A value paired with its “quality” as defined in RFC7231.

ToValues

Token68

A base68 value as defined in RFC7235.

Enums

RetryAfter

Traits

Header

HeaderMapExt

An extension trait adding typed getters and setters to HeaderMap.